+++ src/utils/getClipboardList.ts
@@ -7,7 +7,7 @@ export default async function getClipboards(config: WorkspaceConfiguration): Pro
     if(endpoint.trim().length === 0 || clipboardNamespace.trim().length === 0) return undefined;
     
     try{
-        const connections = await fetch(`${endpoint}/list?namespace=${clipboardNamespace}`);
+        const connections = await fetch(`${endpoint}/list?namespace=${clipboardNamespace}&sort=${config.get<string>("sortResults")!}`);
         if(connections.statusText == "Not Found" && connections.status == 404) return undefined;
         return await connections.json() as string[];
     }catch{